+++ spykeviewer/ui/plugin_model.py
@@ -3,6 +3,7 @@ from PyQt4.QtCore import QAbstractItemModel, QModelIndex
 
 from spykeviewer.plugin_framework.plugin_manager import PluginManager
 
+
 #noinspection PyMethodOverriding
 class PluginModel(PluginManager, QAbstractItemModel):
     """ Implements a Qt-Model for the PluginManager for use in e.g. QTreeView
@@ -114,7 +115,7 @@ class PluginModel(PluginManager, QAbstractItemModel):
 
         return indices
 
-    def get_all_indices(self, parent = None):
+    def get_all_indices(self, parent=None):
         indices = []
         if parent is None:
             parent = self.root
@@ -125,6 +126,20 @@ class PluginModel(PluginManager, QAbstractItemModel):
             indices.append(self.createIndex(i, 0, c))
 
         return indices
+
+    def get_all_folders(self, parent=None):
+        indices = []
+        if parent is None:
+            parent = self.root
+
+        for i in xrange(parent.childCount()):
+            c = parent.child(i)
+            if c.childCount():
+                indices.extend(self.get_all_folders(c))
+                indices.append(self.createIndex(i, 0, c))
+
+        return indices
+
 
     def get_plugins_for_name(self, name):
         """ Return list of plugins with given name